requirements.txt
# Core dependencies for IntelliBase
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
asyncio-extras>=1.3.2

# Data processing
daft>=0.2.0

# Vector database
weaviate-client>=3.24.0

# AI/ML
openai>=1.0.0
arize-phoenix>=1.0.0

# Utilities
python-dotenv>=1.0.0
pathlib2>=2.3.7
typing-extensions>=4.8.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional dependencies for enhanced functionality
jupyter>=1.0.0
notebook>=7.0.0 